moved: feedback, privacy policy, about us pages moved to settings.
